fix: Restore concurrent copy with a fallback to avoid a inner-Copy race condition
Try to copy files with one retry, fail fast and ignore the error if this happens, then retry
copying the files in a second time. This is present to avoid a race condition during
cross-targeted builds when this current target is executed concurrently.

When running, the Copy task may detect that files are not present for two threads/processes, then assume it's safe to copy, while only one of them can do so. Later on, the copied files may be locked by dotnet/msbuild and the other copy instance never finishes sucessfully, regardless of SkipUnchangedFiles.